fix(admin): extract team form client components to fix Vercel pg/dns build error
Admin team pages were marked 'use client' causing Next.js to attempt to bundle
pg (Node.js-only) for the browser via auth.ts -> prisma.ts. Extracted interactive
forms into separate client components; page files are now Server Components.
